Simplify external repo CI config and align tests with style guide
